Use relative time for all Created & Updated timestamps
Extract RelativeTime component to shared location and use it
consistently across the UI:
- Task detail page (Created and Updated times)
- Task detail child tasks table
- Events list
- Event detail page
